fixed pt cut bin range
didn't include the last bin (20-30 GeV) by mistake, so it was added. Also, added a code to state if the fit is from CS or HX, but it was confirmed that it didn't give us difference at least in the full range (-1 to 1 for costheta, -180 to 180 degrees for phi) as expected.